Foseplug_backend/authentication/urls.py
from django.urls import path
from . import views
from .views import MyTokenObtainPairView
from rest_framework_simplejwt.views import (
    TokenRefreshView,
)

app_name="auth"

urlpatterns=[
    path("register/",view=views.register_user,name='registe_user'),
    path('token/', MyTokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
     path('user/', views.getUser, name='getuser'),
    
]