Foseplug_backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
binaryornot==0.4.4
blinker==1.6.2
boto3==1.26.143
botocore==1.29.143
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.2.0
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.0
decorator==5.1.1
distlib==0.3.7
Django==4.2.4
django-browser-reload==1.7.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-fontawesome-5==1.0.18
django-tailwind==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
docxcompose==1.4.0
docxtpl==0.16.7
drf-yasg==1.21.4
dynaconf==3.1.12
filelock==3.12.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
humanize==4.7.0
hvac==1.1.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
knox==0.1.14
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
packaging==23.0
Pillow==9.2.0
platformdirs==3.10.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.6.0
pyPdf==1.6
PyPDF2==2.12.0
pypiwin32==223
pypng==0.20220715.0
python-dateutil==2.8.2
python-docx==0.8.11
python-slugify==8.0.1
pyttsx3==2.90
pytube==12.1.2
pytz==2022.7.1
pywin32==306
PyYAML==6.0
qrcode==7.4.2
redis==5.0.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
segno==1.5.2
shortuuid==1.0.11
sinch==0.1.0
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.3
sympy==1.12
text-unidecode==1.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
virtualenv==20.24.3
Werkzeug==2.3.7
win32-setctime==1.1.0
winshell==0.6
yarl==1.9.2
